vmm: rename sigsys_handler to signal_handler
This step is in preparation to a refactoring that moves generic signal
handling utilities to the sys_util crate, leaving only Firecracker
specifics (read: custom signal handlers themselves) in vmm.
